use home screen as setting screen
