+++ src/screens/home/index.tsx
@@ -1,42 +1,59 @@
 import { StatusBar } from "expo-status-bar";
-import React, { ReactElement } from "react";
+import React, { ReactElement, useState, useEffect } from "react";
+import { StyleSheet, View, Text, TextInput } from "react-native";
+import AsyncStorage from "@react-native-async-storage/async-storage";
 import { StackNavigationProp } from "@react-navigation/stack";
 import { RootStackParamList } from "../../../App";
-import { Button, ThemeProvider, Text } from "react-native-elements";
-import { SafeAreaView } from "react-native-safe-area-context";
 
-const theme = {
-  Button: {
-    containerStyle: {
-      margin: 40,
-    },
-  },
-};
-
-export default function HomeScreen({
-  navigation,
-}: {
-  navigation: HomeScreenNavigationProp;
-}): ReactElement {
+const DEFAULT_ENDPOINT = "127.0.0.1";
+export default function HomeScreen(): ReactElement {
+  const [endpoint, setEndpoint] = useState(DEFAULT_ENDPOINT);
+  useEffect(() => {
+    (async () => {
+      try {
+        const endpointFromPreferences = await AsyncStorage.getItem("@endpoint");
+        if (endpointFromPreferences != null) {
+          setEndpoint(endpointFromPreferences);
+        }
+      } catch (error) {
+        // 設定読み込みエラー
+        console.log(error);
+      }
+    })();
+  }, []);
+  useEffect(() => {
+    (async () => {
+      try {
+        await AsyncStorage.setItem("@endpoint", endpoint);
+      } catch (error) {
+        // 設定保存エラー
+        console.log(error);
+      }
+    })();
+  }, [endpoint]);
+  // TODO サーバーと接続できているか確認するボタンの実装
   return (
-    <ThemeProvider theme={theme}>
-      <Text h1>Card Playroom</Text>
-      <Button
-        title="ルーム作成"
-        onPress={() => navigation.navigate("CreateRoom")}
+    <View style={styles.container}>
+      <Text>Setting Screen</Text>
+      <Text>サーバーアドレス</Text>
+      <TextInput
+        style={{
+          width: "50%",
+          borderWidth: 1,
+          borderColor: "#ccc",
+        }}
+        onChangeText={(input) => setEndpoint(input)}
+        value={endpoint}
       />
-      <Button
-        title="ルーム参加"
-        onPress={() => navigation.navigate("RoomList")}
-      />
-      <Button
-        title="デッキ編集"
-        onPress={() => navigation.navigate("EditDeck")}
-      />
-      <Button title="設定" onPress={() => navigation.navigate("Preferences")} />
       <StatusBar style="auto" />
-    </ThemeProvider>
+    </View>
   );
 }
-
-type HomeScreenNavigationProp = StackNavigationProp<RootStackParamList, "Home">;
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+});
